기능 수정, 기능 추가
- 대미지 공식 수정
물리/마법/고정 대미지 공식에 공통으로 (1+0.01×자신의 현재 분노)식을 추가함

- 일시적 버프 상태일 때 버프 아이콘 표시 추가

+++ Assets/Resources/Scripts/Battle/Skill/Enemy/SingleStrike.cs
@@ -28,7 +28,7 @@ public class SingleStrike : SkillAsset
         yield return originCell; // 단일 대상이므로 원점 셀만
     }
 
-    public override int ComputeDamage(BattleUnit caster, BattleUnit target, in SkillRuntime ctx)
+    public override float ComputeDamage(BattleUnit caster, BattleUnit target, in SkillRuntime ctx)
     {
         if (caster == null || target == null) return 0;
